period
перенесла период в статистику
