+++ src/components/MobileStatisticsNavigation/MobileStatisticsNavigation.js
@@ -3,10 +3,13 @@ import Balance from "../../components/BalanceNavigationSection/Balance";
 import BackToMain from "../../components/BalanceNavigationSection/BackToMain";
 import Period from '../../components/Period';
 
-export default function MobileStatisticsNavigation () {
+export default function MobileStatisticsNavigation ({ handlePreviousPeriod,
+    periodMonth,
+    periodYear,
+    handleNextPeriod}) {
 return (
     <><BackToMain />
-    <Period />
+    <Period handlePreviousPeriod={handlePreviousPeriod} periodMonth={periodMonth} periodYear={periodYear} handleNextPeriod={handleNextPeriod}/>
     <BalanceTitle />
     <Balance /></>
 )
