+++ src/views/StatisticsView/StatisticsView.js
@@ -8,6 +8,10 @@ import useWindowDimensions from "../../hooks/useWindowDimensions";
 import MobileStatisticsNavigation from "../../components/MobileStatisticsNavigation";
 import TabletDesktopStatisticsNavigation from "../../components/TabletDesktopStatisticsNavigation/TabletDesktopStatisticsNavigation";
 import s from "./StatisticsView.module.css";
+import { useState } from 'react';
+import moment from "moment";
+import "moment/locale/ru";
+// import {getMonthlyExpensesThunk, getMonthlyIncomesThunk} from "../../redux/reports/reportsThunk";
 
 // import { useDispatch } from "react-redux";
 // import { useEffect } from "react";
@@ -44,6 +48,41 @@ export default function StatisticsView() {
   //   dispatch(getExpensesThunk())
   // })
   
+// ====added by Y.S.=====
+const date = new Date();
+  // eslint-disable-next-line no-unused-vars
+  const [periodDate, setPeriodDate] = useState(moment(date));
+  const [periodMonth, setPeriodMonth] = useState(moment(date).format('MM'));
+  const [periodYear, setPeriodYear] = useState(moment(date).format('YYYY'));
+  const nextMonth = () => periodDate.add(1, 'month').format('MM');
+  
+  const previousMonth = () => periodDate.subtract(1, 'month').format('MM');
+  const previousYear = () => periodDate.subtract('year').format('YYYY');
+  const nextYear = () => periodDate.add('year').format('YYYY');
+ 
+
+
+const indexOfMonth = (periodMonth) => {
+  periodMonth.forEach((month, idx) =>
+   periodMonth.indexOf(month) === idx);
+   
+}
+
+  const handlePreviousPeriod = () => {
+      setPeriodMonth(previousMonth()); 
+      indexOfMonth  && setPeriodYear(previousYear());
+      
+  };
+
+   const handleNextPeriod = () => {
+     
+      setPeriodMonth(nextMonth());
+     indexOfMonth && setPeriodYear(nextYear());
+      
+  };
+// =====
+
+
   
   const size = useWindowDimensions();
   const { width } = size;
@@ -59,9 +98,9 @@ export default function StatisticsView() {
     <>
       <div className={s.balanceWrapper}>
         {mobileView ? (
-          <MobileStatisticsNavigation />
+          <MobileStatisticsNavigation handlePreviousPeriod={handlePreviousPeriod} periodMonth={periodMonth} periodYear={periodYear} handleNextPeriod={handleNextPeriod}/>
         ) : (
-          <TabletDesktopStatisticsNavigation />
+          <TabletDesktopStatisticsNavigation handlePreviousPeriod={handlePreviousPeriod} periodMonth={periodMonth} periodYear={periodYear} handleNextPeriod={handleNextPeriod}/>
         )}
       </div>
